test(junit): report only the final result for a retried test
A test that passes after `{ retry: N }` was written to JUnit as N
separate <testcase> entries, one <failure> per failed attempt followed
by a passing entry. Every JUnit consumer Bun uploads to (Buildkite Test
Analytics, Jenkins, GitLab) counts each <failure> as a real failure, so
a passing test file with retries shows up red in the dashboard. Bun's
own test/js/bun/test/test-retry-repeats-basic.test.ts was being
reported red on every CI run for this reason.

The per-attempt entries were added in #26866 to give flaky-test
detection per-attempt data, but signalling flakiness via <failure>
inside an otherwise passing file is indistinguishable from a real
failure to any standard JUnit consumer.

Report exactly one <testcase> per test with the final outcome. The
terminal '(attempt N)' suffix already communicates that retries
happened. The FlakyAttempt buffer that recorded intermediate results
has no remaining reader and is removed.

test-retry-repeats-basic.test.ts now spawns its retry/repeats checks in
a subprocess so the intentional intermediate throws stay out of the
parent run's reporter output (no more '::error' annotations for
passing retries in CI logs).
