+++ src/runtime/test_runner/Execution.rs
@@ -152,8 +152,6 @@ pub struct ExecutionSequence {
     pub test_entry: Option<NonNull<ExecutionEntry>>,
     pub remaining_repeat_count: u32,
     pub remaining_retry_count: u32,
-    pub flaky_attempt_count: usize,
-    pub flaky_attempts_buf: [FlakyAttempt; ExecutionSequence::MAX_FLAKY_ATTEMPTS],
     pub result: Result,
     pub executing: bool,
     pub started_at: Timespec,
@@ -164,15 +162,7 @@ pub struct ExecutionSequence {
     pub maybe_skip: bool,
 }
 
-#[derive(Clone, Copy, Default)]
-pub struct FlakyAttempt {
-    pub result: Result,
-    pub elapsed_ns: u64,
-}
-
 impl ExecutionSequence {
-    pub(crate) const MAX_FLAKY_ATTEMPTS: usize = 16;
-
     pub(crate) fn init(
         first_entry: Option<NonNull<ExecutionEntry>>,
         test_entry: Option<NonNull<ExecutionEntry>>,
@@ -186,8 +176,6 @@ impl ExecutionSequence {
             remaining_repeat_count: repeat_count,
             remaining_retry_count: retry_count,
             // defaults:
-            flaky_attempt_count: 0,
-            flaky_attempts_buf: [FlakyAttempt::default(); Self::MAX_FLAKY_ATTEMPTS],
             result: Result::Pending,
             executing: false,
             started_at: Timespec::EPOCH,
@@ -195,10 +183,6 @@ impl ExecutionSequence {
             expect_assertions: ExpectAssertions::NotSet,
             maybe_skip: false,
         }
-    }
-
-    pub(crate) fn flaky_attempts(&self) -> &[FlakyAttempt] {
-        &self.flaky_attempts_buf[0..self.flaky_attempt_count]
     }
 
     fn entry_mode(&self) -> ScopeMode {
@@ -544,18 +528,6 @@ impl Execution {
 
             // Handle retry logic: if test failed and we have retries remaining, retry it
             if test_failed && sequence.remaining_retry_count > 0 {
-                if sequence.flaky_attempt_count < ExecutionSequence::MAX_FLAKY_ATTEMPTS {
-                    let elapsed_ns: u64 = if sequence.started_at.eql(&Timespec::EPOCH) {
-                        0
-                    } else {
-                        sequence.started_at.since_now_force_real_time()
-                    };
-                    sequence.flaky_attempts_buf[sequence.flaky_attempt_count] = FlakyAttempt {
-                        result: sequence.result,
-                        elapsed_ns,
-                    };
-                    sequence.flaky_attempt_count += 1;
-                }
                 sequence.remaining_retry_count -= 1;
                 Execution::reset_sequence(sequence);
                 return;
@@ -744,17 +716,13 @@ impl Execution {
             }
         }
 
-        // Preserve retry/repeat counts and flaky attempt history across reset
-        let saved_flaky_attempt_count = sequence.flaky_attempt_count;
-        let saved_flaky_attempts_buf = sequence.flaky_attempts_buf;
+        // Preserve retry/repeat counts across reset
         *sequence = ExecutionSequence::init(
             sequence.first_entry,
             sequence.test_entry,
             sequence.remaining_retry_count,
             sequence.remaining_repeat_count,
         );
-        sequence.flaky_attempt_count = saved_flaky_attempt_count;
-        sequence.flaky_attempts_buf = saved_flaky_attempts_buf;
 
         // Snapshot counters are keyed by full test name and incremented on every
         // toMatchSnapshot() call. Without this reset, retries / repeats would
